Port to Qt5/CMake build system
* Lua and ZLIB are expecetd to come from system libraries.
  So, lua/ and zlib/ subdirs are completely dropped
* drop most .pro/.pri files, leave some that are not fully ported
* fix all CMakeLists to use modern cmake language and style
* fix bug with QTemporatyFile usage (again)
* fix some warnings in code
* build qtlockedfile, qtsingleapplication, worlded as shared libs
  (only tolua is left as static lib)
* tolua: changes to file layout structure, decrease unneeded nested
  directories depth

TODO: most plugins. But TileZed now starts fine

+++ src/tiled/BuildingEditor/building.cpp
@@ -32,7 +32,7 @@ Building::Building(int width, int height, BuildingTemplate *btemplate) :
     mHeight(height),
     mTiles(TileCount)
 {
-    Q_ASSERT(TileCount == BuildingTemplate::TileCount);
+    Q_ASSERT((int)TileCount == (int)BuildingTemplate::TileCount);
 
     if (btemplate) {
         mTiles = btemplate->tiles();
